wowee

+++ lib/models/location.dart
@@ -8,6 +8,8 @@ class Location {
 
   // convert from firebase doc feild (map) to location object
   Location.fromMap(Map<String, dynamic> map)
-      : latitude = (map ?? const <String, double>{})['latitude'] as double,
-        longitude = (map ?? const <String, double>{})['longitude'] as double;
+      : latitude =
+            (map ?? const <String, double>{})['latitude'].toDouble() as double,
+        longitude =
+            (map ?? const <String, double>{})['longitude'].toDouble() as double;
 }
